+++ docs/changelog.rst
@@ -9,6 +9,9 @@ CHANGELOG
 
 - Fix parser crash after exception during attachment download (refs #4807)
 
+**Documentation**
+
+- Update README.rst
 
 2.116.0    (2025-07-03)
 ----------------------------